feat: improve error handel
