+++ pkg/packet/packet.go
@@ -11,21 +11,6 @@ type Packet struct {
 	payload primitive.Value
 }
 
-// NewError creates a new Packet to represent an error.
-// It takes an error and an optional cause Packet and constructs a Packet with error details.
-func NewError(err error, cause *Packet) *Packet {
-	pairs := []primitive.Value{
-		primitive.NewString("error"),
-		primitive.NewString(err.Error()),
-	}
-
-	if cause != nil {
-		pairs = append(pairs, primitive.NewString("cause"), cause.Payload())
-	}
-
-	return New(primitive.NewMap(pairs...))
-}
-
 // New creates a new Packet with the given payload.
 // It generates a new unique ID for the Packet.
 func New(payload primitive.Value) *Packet {
